Introduce testnet address generation
A simple extension of the API of `generateAddresses` to take in a
protocol magic, allowing testnet address generation.

This was done by simply extending `wallet-wasm` in `js-cardano-wasm` as
it was the simplest solution for now.

This work is  available here for now: https://github.com/rooooooooob/js-cardano-wasm/commit/13a3f49f80841c28e8e358be7f69f79c58e70109 and the depdencies will need to be updated later to reflect this.

As it previously took in JSON for all params the API did not have to
change inside of `android.rs`, `ios.rs`, `rust_native_cardano.h` (or
anything in the `rust/` directory). Same for `android/.../Native.kt`.

This code, however, has not been tested. I lack a Mac to test ios, and
was unable to get the Android build of yoroi-mobile set up locally after
much trouble.

All call-sites within yoroi-mobile will have to be adjusted accordingly
as well. If such an API change is unacceptable due to other users, then
a secondary function for specifying the protocol magic might have to be
created.